Move specs for #expand_path
